Restructured the shared data acess project to make way for the user classes.
