+++ GermanVocabApp.Api.Tests.Unit/Conversion/ItemRequestToDto/ItemRequestConverterTests.cs
@@ -1,6 +1,6 @@
 ﻿using AutoFixture;
 using GermanVocabApp.Api.VocabLists.Models;
-using GermanVocabApp.DataAccess.Shared.DataTransfer;
+using GermanVocabApp.DataAccess.Shared.Vocab.Models;
 
 namespace GermanVocabApp.Api.Tests.Unit.Conversion;
