+++ GermanVocabApp.Api.Tests.Unit/Conversion/ListDtoToResponseConverterTests.cs
@@ -3,7 +3,7 @@ using GermanVocabApp.Api.VocabLists.Conversion.Lists;
 using GermanVocabApp.Api.VocabLists.Models;
 using GermanVocabApp.Core.Contracts;
 using GermanVocabApp.Core.Exceptions;
-using GermanVocabApp.DataAccess.Shared.DataTransfer;
+using GermanVocabApp.DataAccess.Shared.Vocab.Models;
 using Moq;
 
 namespace GermanVocabApp.Api.Tests.Unit.Conversion.ListDtoToResponse;
