+++ GermanVocabApp.Api.Tests.Unit/Conversion/ListRequestToDto/CreateListRequestToDtoConverterTests.cs
@@ -2,7 +2,7 @@
 using GermanVocabApp.Api.VocabLists.Models;
 using GermanVocabApp.Core.Contracts;
 using GermanVocabApp.Core.Exceptions;
-using GermanVocabApp.DataAccess.Shared.DataTransfer;
+using GermanVocabApp.DataAccess.Shared.Vocab.Models;
 using Moq;
 
 namespace GermanVocabApp.Api.Tests.Unit.Conversion;
