+++ GermanVocabApp.Api/VocabLists/Conversion/Items/ItemDtoToResponseConverter.cs
@@ -1,7 +1,7 @@
 ﻿using GermanVocabApp.Api.VocabLists.Models;
 using GermanVocabApp.Core.Contracts;
 using GermanVocabApp.Core.Exceptions;
-using GermanVocabApp.DataAccess.Shared.DataTransfer;
+using GermanVocabApp.DataAccess.Shared.Vocab.Models;
 
 namespace GermanVocabApp.Api.VocabLists.Conversion.Items;
