+++ GermanVocabApp.Api/VocabLists/Conversion/Lists/UpdateListRequestToDtoConverter.cs
@@ -1,6 +1,6 @@
 ﻿using GermanVocabApp.Api.VocabLists.Models;
 using GermanVocabApp.Core.Contracts;
-using GermanVocabApp.DataAccess.Shared.DataTransfer;
+using GermanVocabApp.DataAccess.Shared.Vocab.Models;
 
 namespace GermanVocabApp.Api.VocabLists.Conversion.Lists;
