+++ GermanVocabApp.Api/VocabLists/VocabListsController.cs
@@ -2,8 +2,8 @@
 using GermanVocabApp.Api.VocabLists.Models;
 using GermanVocabApp.Core.Contracts;
 using GermanVocabApp.Core.Exceptions;
-using GermanVocabApp.DataAccess.Shared;
-using GermanVocabApp.DataAccess.Shared.DataTransfer;
+using GermanVocabApp.DataAccess.Shared.Vocab;
+using GermanVocabApp.DataAccess.Shared.Vocab.Models;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System.Net;
