+++ GermanVocabApp.DataAccess.EntityFramework.Tests.Unit/ListRepositoryTests/ListRepositoryAddCommandTests.cs
@@ -1,6 +1,6 @@
 ﻿using GermanVocabApp.DataAccess.EntityFramework.Vocab;
 using GermanVocabApp.DataAccess.EntityFramework.Vocab.Models;
-using GermanVocabApp.DataAccess.Shared.DataTransfer;
+using GermanVocabApp.DataAccess.Shared.Vocab.Models;
 
 namespace GermanVocabApp.DataAccess.EntityFramework.Tests.Unit;
