+++ GermanVocabApp.DataAccess.EntityFramework.Tests.Unit/ListRepositoryTests/ListRepositoryUpdateCommandTests.cs
@@ -2,7 +2,7 @@
 using GermanVocabApp.DataAccess.EntityFramework.Vocab;
 using GermanVocabApp.DataAccess.EntityFramework.Vocab.Conversion;
 using GermanVocabApp.DataAccess.EntityFramework.Vocab.Models;
-using GermanVocabApp.DataAccess.Shared.DataTransfer;
+using GermanVocabApp.DataAccess.Shared.Vocab.Models;
 using Microsoft.EntityFrameworkCore;
 using System.Linq.Expressions;
