+++ GermanVocabApp.DataAccess.EntityFramework/Vocab/Projection/VocabListProjectionExtensions.cs
@@ -1,5 +1,5 @@
 ﻿using GermanVocabApp.DataAccess.EntityFramework.Vocab.Models;
-using GermanVocabApp.DataAccess.Shared.DataTransfer;
+using GermanVocabApp.DataAccess.Shared.Vocab.Models;
 
 namespace GermanVocabApp.DataAccess.EntityFramework.Vocab.Projection;
