--- GermanVocabApp.DataAccess.Shared/IItemRepositoryAsync.cs
@@ -1,9 +1,0 @@
-﻿//using GermanVocabApp.DataAccess.EntityFramework.Models;
-//using GermanVocabApp.DataAccess.Shared.DataTransfer;
-
-//namespace GermanVocabApp.DataAccess.EntityFramework.Repositories;
-//public interface IItemRepositoryAsync
-//{
-//    Task AddRangeAsync(VocabListItemDto[] dtos, Guid listId, DateTime transactionTimeStamp);
-//    void Update(VocabList list, VocabListItemDto[] updateItems, DateTime transactionTimeStamp);
-//}
