--- GermanVocabApp.DataAccess.Shared/IUnitOfWorkAsync.cs
@@ -1,5 +1,0 @@
-﻿namespace GermanVocabApp.DataAccess.Shared;
-public interface IUnitOfWorkAsync
-{
-    Task CommitAsync();
-}
