--- GermanVocabApp.DataAccess.Shared/IVocabUnitOfWorkAsync.cs
@@ -1,7 +1,0 @@
-﻿namespace GermanVocabApp.DataAccess.Shared;
-
-public interface IVocabUnitOfWorkAsync : IUnitOfWorkAsync
-{
-    IVocabListRepositoryAsync ListRepository { get; }
-    IVocabListRepositoryAsync Items { get; }
-}
