+++ GermanVocabApp.DataAccess.Shared/Vocab/Models/Builders/Boilerplate/VocabListDtoBuilder.boilerplate.cs
@@ -1,4 +1,5 @@
-﻿using GermanVocabApp.DataAccess.Shared.DataTransfer;
+﻿using GermanVocabApp.DataAccess.Shared.Core;
+using GermanVocabApp.DataAccess.Shared.Vocab.Models;
 
 namespace GermanVocabApp.DataAccess.Models.Builders;
